figure-caption-extraction/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
transformers
torch
